# Implement TakeLast performance optimizations and benchmarks