#204 
Simple TakeLast(100) Benchmark.
I wonder if optimization by TryGetSpan is needed.
(In fact, it might have been better if ValueEnumerable had an InnerSkip functionality)
| Method                  | N       | Mean           | Error          | StdDev       | Gen0   | Allocated |
|------------------------ |-------- |---------------:|---------------:|-------------:|-------:|----------:|
| FromArray_Linq          | 1000000 |       228.8 ns |       33.15 ns |      1.82 ns | 0.0086 |     136 B |
| FromArray_ZLinq         | 1000000 | 3,547,090.8 ns |  171,427.16 ns |  9,396.50 ns |      - |      48 B |
| FromArray_ZLinq(PR)     | 1000000 |   469,546.8 ns |  162,486.33 ns |  8,906.42 ns |      - |         - |
|                         |         |                |                |              |        |           |
| FromList_Linq           | 1000000 |       242.9 ns |       28.69 ns |      1.57 ns | 0.0086 |     136 B |
| FromList_ZLinq          | 1000000 | 3,547,946.9 ns |  122,781.05 ns |  6,730.04 ns |      - |      48 B |
| FromList_ZLinq(PR)      | 1000000 |   545,908.3 ns |    4,946.87 ns |    271.15 ns |      - |         - |
|                         |         |                |                |              |        |           |
| FromEnumerable_Linq     | 1000000 | 4,456,278.1 ns |   76,890.24 ns |  4,214.61 ns |      - |    1368 B |
| FromEnumerable_ZLinq    | 1000000 | 4,021,346.6 ns |  164,733.85 ns |  9,029.62 ns |      - |     136 B |
| FromEnumerable_ZLinq(PR)| 1000000 | 3,410,969.7 ns |1,775,177.79 ns | 97,303.49 ns |      - |     128 B |