@Akeit0 2025-07-18
Hmm... How should the test be passed?
[BinaryOperations_SourceEnumeratorsDisposed](https://github.com/Cysharp/ZLinq/blob/1e77b7e08f4091d9e71f01fed155bc0d2922dbfa/tests/System.Linq.Tests/Tests/ZLinq/LifecycleTests.cs#L50)
I don't think this is a good way.
```csharp
var totalCount = 0X7FFFFFC7;//Array.MaxLength
if (0X7FFFFFC7 < takeCount||source.TryGetNonEnumeratedCount(out totalCount))
```