@neuecc 2025-07-25
Thank you, I think it's very good.
And I see, there's such a compat specification...!
Yes, it seems better to carry it over, but maybe another time.